requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.2
altair==5.4.1
appdirs==1.4.4
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==24.2.0
babel==2.16.0
bio-embeddings==0.2.2
bio-embeddings-allennlp==0.9.2
bio-embeddings-bepler==0.0.1
bio-embeddings-CPCProt==0.0.1
bio-embeddings-deepblast==0.1.0
bio-embeddings-plus==0.1.1
bio-embeddings-tape-proteins==0.5
biopython==1.83
blinker==1.8.2
blis==0.7.11
boto3==1.35.21
botocore==1.35.21
cachetools==5.5.0
catalogue==1.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
conllu==4.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
docutils==0.20.1
editdistance==0.8.1
exceptiongroup==1.2.2
fair-esm==0.4.0
filelock==3.16.1
flaky==3.8.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
gensim==3.8.3
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
h5py==3.11.0
huggingface-hub==0.25.0
humanize==3.14.0
idna==3.10
imagesize==1.4.1
importlib-metadata==4.13.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.1.77
jax-unirep==1.0.3
jaxlib==0.4.13
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonnet==0.20.0
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lightning-utilities==0.11.7
llvmlite==0.41.1
lmdb==1.5.1
lock==2018.3.25.2110
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.1.0
multipledispatch==1.0.0
murmurhash==1.0.10
narwhals==1.8.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
numpydoc==1.7.0
opt-einsum==3.3.0
optuna==4.0.0
overrides==4.1.2
packaging==24.1
pandas==1.5.3
parsimonious==0.10.0
pillow==10.4.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
plotly==5.24.1
pluggy==1.5.0
preshed==3.0.9
protobuf==5.28.1
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-slugify==5.0.2
pytorch-lightning==1.8.6
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
responses==0.25.3
rich==13.8.1
rpds-py==0.20.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==0.24.2
scipy==1.10.1
sentencepiece==0.2.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
snowballstemmer==2.2.0
spacy==2.3.9
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.35
sqlparse==0.5.1
srsly==1.0.7
streamlit==1.38.0
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==7.4.6
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==1.9.1
torchmetrics==1.2.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing-utils==0.1.0
typing_extensions==4.12.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==1.26.20
wasabi==0.10.1
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.4
word2number==1.1
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.2
zope.event==5.0
zope.interface==7.0.3
